refactor: Rename handleBurgerButton to mobileMenu
